Travel Time solved

+++ src/main/java/TravelTime.java
@@ -2,5 +2,5 @@
 @FunctionalInterface
 public interface TravelTime {
 
-    double calculateTime(int dist,int avg);
+    int calculateTime(int dist,int avg);
 }
